requirements.txt
django
pymongo
python-dotenv
# python-reloader
bcrypt
djangorestframework
django-rest-framework-mongoengine
mongoengine
pillow
gunicorn
django-cors-headers
python-dotenv
cloudinary